[refactor] provide `measure` in `Reusable`
