+++ server/invoker/index.test.js
@@ -22,42 +22,6 @@ describe("Test Invoker", () => {
   });
 });
 
-describe("Test ReusableInvoker class", () => {
-  class SampleInvoker extends ReusableInvoker {
-    constructor() {
-      super();
-    }
-
-    async _invoke() {
-      return { result: "hello", elapsed: 16 };
-    }
-  }
-
-  test("Invoker can invoke task", async () => {
-    const invoker = new SampleInvoker();
-    const { result, elapsed } = await invoker._invoke();
-    expect(result).toBe("hello");
-    expect(elapsed).toBe(16);
-  });
-
-  test("Invoker can run", async () => {
-    const invoker = new SampleInvoker();
-    const result = await invoker.run();
-    expect(result).toBe("hello");
-  });
-
-  test("Invoker can be reused and get average elapsed time", async () => {
-    const invoker = new SampleInvoker();
-    for (let i = 0; i < 10; i += 1) {
-      const result = await invoker.run();
-      expect(result).toBe("hello");
-    }
-    const avgElapsed = invoker.averageElapsedTime;
-    // approximately equals to 16 with 2 significant figures
-    expect(avgElapsed).toBeCloseTo(16);
-  });
-});
-
 describe("Test SwitchingInvoker", () => {
   const name = "hoge";
   const expectedResult = expect.stringContaining(`hello, ${name}`);
@@ -85,7 +49,7 @@ describe("Test SwitchingInvoker", () => {
       const result = await invoker.run(name);
       expect(result).toEqual(expectedResult);
     }
-    const avgElapsed = invoker.averageElapsedTime;
+    const avgElapsed = invoker.averageElapsedTime();
     expect(avgElapsed).toBeGreaterThan(0);
   });
 });
